requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework==3.9.2
django-cors-headers==2.5.2
ipdb
ipython